backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
python-dotenv>=1.0.0
anthropic>=0.40.0
websockets>=12.0
opencv-python>=4.9.0
mediapipe>=0.10.9
